MTSDK-381_TransportSDK_can_consume_structured_messages_from_backend
-update gradle wrapper files
-make suggested changes
-add to unit test
