+++ transport/src/androidUnitTest/kotlin/transport/core/MessageTest.kt
@@ -15,7 +15,6 @@ import com.genesys.cloud.messenger.transport.shyrka.WebMessagingJson
 import com.genesys.cloud.messenger.transport.utility.AttachmentValues
 import com.genesys.cloud.messenger.transport.utility.MessageValues
 import com.genesys.cloud.messenger.transport.utility.QuickReplyTestValues
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import org.junit.Test
 
@@ -42,6 +41,7 @@ class MessageTest {
             assertThat(attachments).isEmpty()
             assertThat(events).isEmpty()
             assertThat(quickReplies).isEmpty()
+            assertThat(cards).isEmpty()
             from.run {
                 assertThat(this).isEqualTo(expectedParticipant)
                 assertThat(name).isNull()
