--- transport/src/commonMain/kotlin/com/genesys/cloud/messenger/transport/core/Carousel.kt
@@ -1,8 +1,0 @@
-package com.genesys.cloud.messenger.transport.core
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Carousel(
-    val cards: List<Card>
-)
